implemented functionality to store data tied up with k8s cluster creation into DB
